Update how display and ordinal values are returned
Update the way the `entry_maker` function builds the `display` and
`ordinal` values to be returned for each entry, as I found that a lot of
times Telescope would not show any results until the user started typing
in the search box.

I tried using both `table.concat` as well as the string concatenation
operator (`..`) in the table returned by this function and found that,
only when using `table.concat` and assigning to a variable, could I get
the results to consistently appear in the results list.